Move files into the src directory
