+++ package.json
@@ -4,7 +4,7 @@
   "description": "Find out who plays the game you are excited about!",
   "main": "bot.js",
   "scripts": {
-    "start": "node bot.js",
+    "start": "node src/bot.js",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "ellisande",
